change: moved authentication route to new route folder
